### 2019-04-01 #1361 @whedon
[REVIEW]: SMACT: Semiconducting Materials by Analogy and Chemical Theory
**Submitting author:** @dandavies99 (<a href="http://orcid.org/0000-0003-4094-5992">D. W. Davies</a>)
**Repository:** <a href="https://github.com/WMD-group/SMACT/" target ="_blank">https://github.com/WMD-group/SMACT/</a>
**Version:** v2.1
