### 2019-04-02 #1361 @symmy596
I am happy and indeed keen to review SMACT.

However I would like to flag that I know the lead author Dan as we have worked in the same department (Bath University) in the past, although he has since moved on. With that said we have had no collaborations, have no current plans for future collaborations. I feel there is no conflict of interest, however I do want to flag it now in the interest of transparency.
