### 2019-04-02 #1361 @symmy596
Update - Dan left Bath in October.
