### 2019-04-02 #1361 @danielskatz
As an associate editor in chief, I don't think this conflict should prevent you from reviewing the submission, but we do greatly appreciate you mentioning and recording it.
