### 2019-04-15 #1361 @symmy596
I would like to start by saying that SMACT is a really nice piece of work that has a clear research use. The authors have done a great job.

### Reviewer comments

#### General Comments

- The README on PyPi has not rendered correctly. PyPi expects readme files to be rst files however you can get away with markdown by putting - long_description_content_type=’text/markdown into your setup.py. See https://github.com/symmy596/SurfinPy/blob/master/setup.py for an example.

- It is not clear from the README what to do if you have any general quaetions that may not be suited to the issue tracker. It needs to be made clear in the README how a user can contact the developers with questions. This falls under point 3 of the community guidelines section of the review criteria.

- Two of the repo badges have not rendered correctly.

- There are automated tests however there is no mention of them in the README. A sentence explaining that tests exist and an explanation of how to run them is needed.

#### Examples

The examples work as expected, however I feel that some of the descriptions within the notebooks are a bit thin and could be fleshed out to provide more detail. Some of your future users may be completely new to this field and anything that you can do in your examples to help them get started should be encouraged.

- There is no real background provided in the notebooks. In the Inverse_formate_perovskite example it is clear what you are doing but there is no real explanation of why you are doing it. Someone who comes across SMACT is more likely to start using it if they can immediately see how it can fit into their research. A sentence or two that explains why an example exists helps a user to see how they could use it for their own research.

- The "next step" is missing in some examples. For example in the cation_mutation example how could a user take the information and convert it into something useful e.g. a POSCAR.

- The documentation does a good job at explaining the workflow. In the docs you point to the Generate_compositons_list.ipynb and say that this is an example of generating a pandas dataframe which could be used as an input for ML. It would be useful to users if you provided an example of how to actually do that, even if this was just a code block within the markdown (As I guess it could be computationally intensive).

- It is reasonable to expect a certain level of background knowledge; however as these are tutorials I think it is better to provide more background than you think you need. For example, in Inverse_formate_perovskites you mention the electronegativity order test and the Goldschmidt tolerance factor. It may be worth providing definitions of these terms (and others like it in other examples) in order to help users who have not come across these terms before.

- In the Inverse_formate_perovskites tutorial you mention that the methodology can be carried out in fewer lines of code but that the method has been broken down for clarity. This is a great idea from a learning standpoint however as a user will typically be using the "fewer lines of code" method I think that method should be provided. It could be put into a code block within the markdown.
