### 2019-04-17 defa4bc
Update README and setup.py to clarify developer contact, automated tests and remove broken badges
