### 2019-04-18 5d38f47
Add explanations to InversePerovskites example and fix bug in __init__ for how shannon radii are accessed
