+++ examples/InversePerovskites/README.md
@@ -1,2 +1,6 @@
- This example searches for charge inverted perovskites, i.e. with an anion on the A site.
- It assumes a formate moelcule as the C-site and uses Goldschmidt ratios as part of the screening.
+ Perovskites are an important class of inorganic materials, particularly for solar energy applications.
+ Usually, the perovskite structure contains an A cationm, a B cation and an X anion in the ratio 1:1:3 ([see wikipeidia for more information](https://en.wikipedia.org/wiki/Perovskite_(structure))).
+ Here we search for charge inverted perovskites, i.e. with an anion on the A site. This class of material is closely related to perovskites, and may represent another fruitful search space for new photovoltaic materials. 
+
+In this example  we assume a simple [formate moelcule](https://en.wikipedia.org/wiki/Formate) as the C-site and uses [Goldschmidt ratio rules](https://en.wikipedia.org/wiki/Goldschmidt_tolerance_factor) as part of the screening.
+We also apply the standard charge neutrality and electronegativity tests as described [in the docs](https://smact.readthedocs.io/en/latest/examples.html#neutral-combinations).
