### 2019-04-18 25c8a0a
Add ionic radius to species properties
