### 2019-04-18 7c242e1
Provide more information and context in examples, as well as links to other resources
