+++ examples/CationMutation/README.md
@@ -1,3 +1,7 @@
+The concept of cation mutation (ion substitution) to tune the properties of compounds has been around at least [since the 1950s](https://www.sciencedirect.com/science/article/pii/0022369758900507) and has proved a useful concept in [more recent studies](https://journals.aps.org/prb/abstract/10.1103/PhysRevB.79.165211).
+
+![](cation_mutation.png)
+
 This example generates all possible (symmetry inequivalent) substitutions of Sr on Ba sites in the cubic perovskite
-BaTiO3; single and double substitutions. 
-The distorter module uses the ASE Python library to achieve this. 
+BaTiO3; single and double substitutions.
+The distorter module uses the [ASE Python library](https://wiki.fysik.dtu.dk/ase/) to achieve this.
