### 2019-04-18 #1361 @dandavies99
Thanks for the really useful feedback @symmy596 and with many helpful tips. We have tried to address all your points as best we can through changes to the repository.

**General**

- We have added the `long_description_content_type` flag to get the README to render correctly on PyPi for the next release.

- We have added developer contact details of to the [README](https://github.com/WMD-group/SMACT/blob/master/README.md#development-notes).

- It appears that since submitting the repo to JOSS, two of the badge providers had developed problems ([Hitcount](http://hits.dwyl.io/) and [Badginator](https://github.com/defunctzombie/badginator)). The former is working again. As fun as the Badginator badge is, it is in no way essential so we have just removed it. 🤷‍♂️

- Regarding automated tests: There were already instructions under `Code contributions` on how to run them, but we have now made this more obvious by adding a [separate section](https://github.com/WMD-group/SMACT/blob/master/README.md#tests).

**Examples**

Thanks for looking at these in detail and for the useful comments. We have worked on the accessibility of all the examples including:

- We have fleshed out the READMEs for the example notebooks to provide some extra background, images and links to resources for defining key terms (e.g. inverse formate perovskite [example](https://github.com/WMD-group/SMACT/blob/master/examples/InversePerovskites/Inverse_formate_perovskites.ipynb)).

- The cation mutation [example](https://github.com/WMD-group/SMACT/tree/master/examples/CationMutation) now includes a demo of how the structure objects can be written out to a cif file.

- The [Generate_compositions_lists example](https://github.com/WMD-group/SMACT/blob/master/examples/Counting/Generate_compositions_lists.ipynb) now includes a **Next steps** section which has a code block demonstrating featurisation for machine learning applications. Instead of including an actual machine learning section -- which would be quite long and isn't really the focus of SMACT -- we have provided a link to a [full example in another repo](https://github.com/WMD-group/Solar_oxides_data) for the interested user.

- We now demonstrate both the long (beginner) and short (advanced)  approaches fully in the [inverse formate preovskite example](https://github.com/WMD-group/SMACT/blob/master/examples/InversePerovskites/Inverse_formate_perovskites.ipynb). This also revealed a bug regarding how ionic radii are read in so that has also been fixed!
