### 2019-05-12 #1361 @utf
I have finished reviewing the paper and the repository. Overall, SMACT is a clean and well documented package that has many useful features. It will be of use to many researchers in the materials informatics and materials design communities.

I have a number of suggestions for improving the package.

#### Documentation and examples

- The readme could be adjusted to make the functionality of the package more obvious. The statement of need is rather broad and not too revealing, whereas the contents section details what the code components do but not how they fit together. Overall, there is a lot of functionality which is often not obviously related (e.g., the distortion module). In my opinion, the readme would be better if it a) clearly listed the features of the code and b) gave specific use cases with corresponding links to examples (and maybe a paper).
- The readme would benefit from a more prominent link to the online documentation. I actually had to search the page for the word "doc" to find it.
- I found the examples page in the online docs to give the clearest explanation of the functionality of the code. I would direct users there first and suggest reading the other examples afterwards.
- There are examples in both the package repository and a separate repository. The separation doesn't seem entirely clear to me. I would consolidate the examples to avoid confusion. On that note, the readme lists both example locations but the [Introduction page](https://smact.readthedocs.io/en/latest/introduction.html) of the online docs just lists the separate repo.
- The [Requirements section](https://smact.readthedocs.io/en/latest/getting_started.html#requirements) of the getting started page is a bit funky.
- There is a typo in the "Inverse_formate_perovskites.ipynb" example: "senseibe" (or maybe this is a new type of perovskite?)

#### SMACT package

These are more general comments that do not necessarily need to be addressed but which occurred to me during the review.

- Personally I find `smact_test` a bad function name. It isn't clear what the function does from the name alone and actually suggests it is something to do with testing. Could this be called `smact_filter` instead?
- In many of the examples, I see the paradigm:
```python
for element in search:
    with open(os.path.join(data_directory, 'shannon_radii.csv'),'r') as f:
        reader = csv.reader(f)
```
As you are often looping over ~100 elements, to constantly reload the file seems excessive. It would be better if all the additional data files were just json dictionaries. The data could be loaded once and then just accessed using `shannon_radii[el][coordination]`.

Having the data files as dictionaries might also remove the need for the complex `data_loader` caching functionality. The data folder is only 200 KB, so all the tables could be loaded into memory when the package is imported (in `__init__.py`). The data would then just be module level dictionaries that are accessed by all element/species objects.
