### 2019-05-22 68f8df0
Continue improving README
