### 2019-05-22 e5fc790
Merge all examples into examples directory
