+++ examples/Practical_tutorial/Site/surface_points.py
@@ -1,0 +1,313 @@
+#!/usr/bin/env python 
+# Defines the positions (on a 2D grid) of the positions of under-coordinated
+# atoms at the surface of a material
+################################################################################
+# Copyright Keith T Butler    (2015)		                               #
+#                                                                              #
+# This file is part of SMACT: builder.py is free software: you can             #
+# redistribute it and/or modify it under the terms of the GNU General Public   #
+# License as published by the Free Software Foundation, either version 3 of    #
+# the License, or (at your option) any later version.                          #
+# This program is distributed in the hope that it will be useful, but WITHOUT  #
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or        #
+# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for    #
+# more details.                                                                #
+# You should have received a copy of the GNU General Public License along with #
+# this program.  If not, see <http://www.gnu.org/licenses/>.                   #
+#                                                                              #
+################################################################################
+
+'''
+The general form of the surface_points file.
+
+Each new species or class of materials is defined as a function, the function contains a dictionary of the known surfaces of that material or class.
+The surfaces contain the 2D projection of the under-coordinated surface atoms, there may be several possible terminations, then there are all listed.
+
+The function has a list called 'exists', this contains all surfaces for which surface points are available. If you add a new surface be sure to include it in this list too.
+
+Each dictionary item contains a list of the surface cuts and must end with an empty tuple, note the ,() at the end of each item.
+'''
+
+def anatase(miller):
+    exists = ['001','010','110','100','101']
+    if miller in exists:
+        surfaces = {}
+        surfaces['110'] = ([0.,0.],[0.2,0],[0.3,0.5],[0.5,0.5],[0.7,0.5],[0.8,0.]),()
+        surfaces['001'] = ([0.5,0.],[0.5,0.5]),()
+        surfaces['010'] = ([0.5,0.],[0.,0.],[0.,0.2],[0.5,0.5],[0.5,0.75],[0.8,0.]) ,()
+        surfaces['100'] = ([0.5,0.],[0.,0.],[0.,0.2],[0.5,0.5],[0.5,0.75],[0.8,0.]) ,()
+        surfaces['101'] = ([0.,0.17],[0.5,0.12],[0.,0.37],[0.,0.62],[0.5,0.67],[0.5,0.87]),([0.,0.14],[0.,0.34],[0.5,0.68],[0.5,0.84]),([0.,0.15],[0.,0.36],[0.5,0.4],[0.,0.57],[0.5,0.65],[0.5,0.86],[0.,0.9]),()
+
+        return surfaces[miller]
+    else:
+	print "No non-polar surface",miller,"is currently in the database, maybe you want to add it."
+	return  []
+
+def TiO2a(miller):
[…]
