+++ examples/Practical_tutorial/Counting/element_data.txt
@@ -1,0 +1,159 @@
+###############################################################################
+#                                                                             #
+#           element_data.txt -- This data file is part of SMACT               #
+#                                                                             #
+###############################################################################
+#                                                                             #
+# This table provides elemental properties which are accessed by the Element  #
+# class. For efficiency the contents are cached in memory the first time they #
+# are accessed by smact.data_loader.                                          #
+#                                                                             #
+# Symbol: Standard chemical symbol used as index throughout SMACT             #
+#                                                                             #
+# Name:   Standard chemical name of element                                   #
+#                                                                             #
+# Z:      Proton number                                                       #
+#                                                                             #
+# Mass:   Relative atomic mass in atomic mass units (i.e. 1/12 * mass of C-12)#
+#         Values obtained from NIST Standard Reference Database 144           #
+#         (J. S. Coursey, D. J. Schwab, J. J. Tsai, NIST Physical Measurement #
+#         Laboratory) in 2016. Where the relative abundance of isotopes is    #
+#         unknown or a range of values is provided, a simple mean was taken.  #
+#                                                                             #
+# r_cov:  Covalent radius in Angstroms                                        #
+#         Ref: Cordero et al. (2008) DOI:10.1039/B801115J                     #
+#                                                                             #
+# e_affinity: Electron affinity in eV. NIST-recommended values.               #
+#         Note that unspecified negative values (originally reported as '<0') #
+#         are listed as 'None' in this table. Electron affinities of elements #
+#         Fr (Z=87) and above are unknown, and also listed as 'None'.         #
+#         Ref: Andersen et al. (1999) DOI: 10.1063/1.556047                   #
+#                                                                             #
+# p_eig:  Highest occupied p-state eigenvalue in eV. Based on approximate     #
+#         Hartree-Fock calculations published by Herman & Skillman (1963);    #
+#         the relevant values were tabulated and converted to eV by Harrison  #
+#         (1980). Harrison's values are used here.                            #
+#                                                                             #
+# s_eig:  Highest occupied s-state eigenvalue in eV. The source is the same   #
+#         as for p_states.                                                    #
+#                                                                             #
+# Abundance: Crustal abundance in mg/kg.                                      #
+#         Ref: CRC Handbook of Chemistry and Physics, 92nd Ed. (2011)         #
+#                                                                             #
+# el_neg: Pauling electronegativity                                           #
+#         Ref: CRC Handbook of Chemistry and Physics, 92nd Ed. (2011)         #
[…]
