+++ examples/SimpleWrappers/README.md
@@ -1,2 +1,0 @@
-This notebook gives a simple example of how SMACT can be incorporated into a python script 
-for use at the command line.

[…]
