### 2019-05-22 615180a
Fix typos in docs and examples
