+++ docs/getting_started.rst
@@ -6,9 +6,7 @@ Getting Started
 Requirements
 ============
 
-
 The main language is Python 3 and has been tested using Python 3.6+. Basic requirements are Numpy and Scipy.
-spglib, and pymatgen are also required for many components.
 The `Atomic Simulation Environment <https://wiki.fysik.dtu.dk/ase>`_
 (ASE), `spglib <http://atztogo.github.io/spglib>`_,
 and `pymatgen <http://pymatgen.org>`_ are also required for many components.
@@ -31,13 +29,13 @@ Alternatively, the latest master branch from the Git repo can be installed using
 
 Then ensure that the location of :mod:`smact` is on your PYTHONPATH.
 
-For developer installation SMACT can be installed from a copy of the source repository (https://github.com/wmd-group/smact); 
+For developer installation SMACT can be installed from a copy of the source repository (https://github.com/wmd-group/smact);
 this will be preferred if using experimental code branches.
 
-To clone the project from Github and make a local installation: 
+To clone the project from Github and make a local installation:
 
 .. code::
-    
+
     git clone https://github.com/wmd-group/smact.git
     cd smact
     pip install --user -e .
