### 2019-05-23 e39c701
Change name of smact_test function to smact_filter
