### 2019-05-23 #1361 @dandavies99
Thank you both @symmy596 and @utf for your time.

@utf, thanks for the the thorough review and useful comments. We have tried to address all the points by making these changes:

- We agree that the separation between the examples in the repo and the workflows in the separate workflow is actually not that helpful. We have moved all examples into the examples folder in the repo and provided a [separate README](https://github.com/WMD-group/SMACT/tree/master/examples) there to signpost to each example.

- We have re-worked the [main README](https://github.com/WMD-group/SMACT/blob/master/README.md):
  - providing an obvious link to the docs near the top of the page, as well as links to examples.
  - We have changed the old “Usage” section to  a “Getting started” section, again signposting the docs as a first port of call, and that further examples can be found in the exampled directory.
  - We have provided a more in-depth list of “Code features” with links to examples and publications, as suggested.

- We have brought the [Requirements section of the docs](https://smact.readthedocs.io/en/latest/getting_started.html#requirements) up to date with the Requirements in the README and fixed the (very well-spotted) ‘sensible’ typo. Sadly, not a new type of perovskite 😞

- We updated the matminer  StrToComposition featurizer in the solaroxides and counting examples.

- We have changed the `smact_test` function to `smact_filter` in the [screening module](https://github.com/WMD-group/SMACT/blob/master/smact/screening.py) (and in all relevant examples and tests), as we agree this better reflects its purpose. We chose this over anything to do with charge_balance as the charge neutrality and electronegativity order filters may be applied independently if desired.

- We agree that continually reopening files within loops should be avoided. We have kept the paradigm as it is for now in the [formate perovskite example](https://github.com/WMD-group/SMACT/tree/master/examples/Inverse_perovskites) in an attempt to make it clear where exactly the data is coming from. We have, however, also added a note and pseudo-code snippet explaining how this should be done for larger production runs.

- Thank you for the great idea regarding loading the data into memory when the package is imported. This is a discussion that some of the developers have been having recently and it is likely that we will simplify the way in which data is loaded in future versions of SMACT.
