### 2019-05-23 #1361 @utf
Hi @dandavies99 the new readme is a lot clearer and is very easy to follow. Really nice work.

@danielskatz I’m happy that all my comments have been thoroughly addressed.
