### 2019-06-09 #1361 @usethedata
@dandavies99 Yes, please go ahead and create the Zenodo archive.  Post the DOI here when you're done.

Sorry for the delays on my part.  I was back on 6/1, but getting dug out from under took longer than I'd hoped.
