### 2019-06-10 #1361 @labarba
Before you make the archive, @dandavies99, we're going to need a new tagged release of the software repository, after all revisions.

- [ ] Please list the latest version tag here, to update the metadata
- [ ] Check the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who made a minor commit are not on it)
- [ ] Please list the Zenodo DOI of the archived version here.
