### 2019-06-10 4892a45
update setup.py to latest release
