### 2019-06-10 #1361 @dandavies99
No problem, @usethedata.
Thanks, @labarba.
- The latest version is v2.1
- This has now been archived on Zenodo with the correct metadata --> [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3242315.svg)](https://doi.org/10.5281/zenodo.3242315)
