### 2019-06-10 f28c776
spell out link URL instead of link on word "here"
[style] word "here" makes no sense when reading the paper: use sensical parts of speech on hyperlinks
(also added backticks to module name)
