### 2019-06-10 #1361 @labarba
@dandavies99 : could you merge my stylistic fix here?
https://github.com/WMD-group/SMACT/pull/20
