### 2019-06-10 #1361 @labarba
Congratulations, @dandavies99, your JOSS paper is published!

Big thanks to our editor: @usethedata, and reviewers: @symmy596, @utf — we appreciate your contribution to JOSS 🙏
